IDs as u16.

+++ src/ids.rs
@@ -2,20 +2,20 @@
 
 pub const TITLE:            &str = "ID3 Padding Remover";
 
-pub const ICO_FROG:         i32 = 101;
+pub const ICO_FROG:         u16 = 101;
 
-pub const MNU_MAIN:         i32 = 200;
-pub const MNU_FILE_OPEN:    i32 = 201;
-pub const MNU_FILE_EXCSEL:  i32 = 202;
-pub const MNU_FILE_REMPAD:  i32 = 203;
-pub const MNU_FILE_REMART:  i32 = 204;
-pub const MNU_FILE_REMRG:   i32 = 205;
-pub const MNU_FILE_PRXYEAR: i32 = 206;
-pub const MNU_FILE_CLRDIAC: i32 = 207;
-pub const MNU_FILE_ABOUT:   i32 = 208;
+pub const MNU_MAIN:         u16 = 200;
+pub const MNU_FILE_OPEN:    u16 = 201;
+pub const MNU_FILE_EXCSEL:  u16 = 202;
+pub const MNU_FILE_REMPAD:  u16 = 203;
+pub const MNU_FILE_REMART:  u16 = 204;
+pub const MNU_FILE_REMRG:   u16 = 205;
+pub const MNU_FILE_PRXYEAR: u16 = 206;
+pub const MNU_FILE_CLRDIAC: u16 = 207;
+pub const MNU_FILE_ABOUT:   u16 = 208;
 
-pub const ACT_MAIN:         i32 = 300;
+pub const ACT_MAIN:         u16 = 300;
 
-pub const DLG_MAIN:         i32 = 1000;
-pub const LST_FILES:        i32 = 1001;
-pub const LST_FRAMES:       i32 = 1002;
+pub const DLG_MAIN:         u16 = 1000;
+pub const LST_FILES:        u16 = 1001;
+pub const LST_FRAMES:       u16 = 1002;
